chore(reorg): Restructure components directory
