--- components/logo.tsx
@@ -1,9 +1,0 @@
-import { c } from '../lib/utils';
-import styles from '../styles/Logo.module.scss';
-
-export default function Logo() {
-  return (
-    <span className={c(styles.logo)}>S</span>
-  );
-}
-
