+++ components/nav/logo.tsx
@@ -1,0 +1,9 @@
+import { classNames as c } from '../../lib/utils';
+import styles from '../../styles/sections/Logo.module.scss';
+
+export default function Logo() {
+  return (
+    <span className={c(styles.logo)}>S</span>
+  );
+}
+
