+++ docker-compose.yml
@@ -1,71 +1,8 @@
-# vim: ts=2:sts=2:sw=2:expandtab
-
-version: '3.7'
-
 services:
   serubin-net:
     image: ${IMAGE_TAG} # Filled by docker ci
     build: .
-    deploy:
-      replicas: 1
-      restart_policy:
-        condition: on-failure
-    restart: always
+    ports:
+      - "3001:80"
     networks:
       - default
-      - traefik-public
-    labels:
-      - "traefik.enable=true"
-      - "traefik.domain=serubin.net"
-      - "traefik.frontend.rule=Host: serubin.net, www.serubin.net, serubin.com, www.serubin.com, serub.in, www.serub.in, solomonrub.in, www.solomonrub.in, solomonrubin.com, www.solomonrubin.com"
-      - "traefik.tags=traefik-public"
-      - "traefik.docker.network=traefik-public"
-      - "traefik.port=80"
-      - "traefik.webservice.frontend.entryPoints=http,https"
-      - "traefik.default.protocol=http"
-
-#  ghost:
-#    image: ghost:alpine
-#    restart: always
-#    deploy:
-#      replicas: 1
-#      restart_policy:
-#        condition: on-failure
-#    environment:
-#      database__client: mysql
-#      database__connection__host: db
-#      database__connection__user: root
-#      database__connection__password: $DB_PASSWD
-#      database__connection__database: ghost
-#      url: https://serubin.net/blog-test
-#    volumes:
-#      - /opt/data/serubin-net/blog/content:/var/lib/ghost/content
-#    networks:
-#      - default
-#      - traefik-public
-#    labels:
-#      - "traefik.enable=true"
-#      - "traefik.domain=serubin.net"
-#      - "traefik.frontend.rule=Host: serubin.net, www.serubin.net, www.serubin.com, serub.in, www.serub.in, solomonrub.in, www.solomonrub.in, solomonrubin.com, www.solomonrubin.com; PathPrefix: /blog-test"
-#      - "traefik.tags=traefik-public"
-#      - "traefik.docker.network=traefik-public"
-#      - "traefik.port=2368"
-#      - "traefik.default.protocol=http"
-#
-#  db:
-#    image: mysql:5.7
-#    restart: always
-#    deploy:
-#      replicas: 1
-#      restart_policy:
-#        condition: on-failure
-#    environment:
-#      MYSQL_DATABASE: ghost
-#      MYSQL_ROOT_PASSWORD: $DB_PASSWD
-#    volumes:
-#      - /opt/data/serubin-net/blog/db:/var/lib/mysql
-#
-networks:
-  traefik-public:
-    external: true
-
