+++ lib/utils.ts
@@ -1,8 +1,3 @@
-
-export const c = (cls: string | string[]) => {
-  if (Array.isArray(cls)) {
-    return cls.join(" ");
-  }
-
-  return cls;
-};
+export const classNames = (...classes: string[]) => {
+  return classes.join(" ");
+}
